Rework documentation entirely
Remove old documentation

New documentation hacks so that it builds for pre-OTP27

It uses a parse transform that removes all documentation attributes from
the module. This is so that in the future, the diff removing this will
be smaller.

Note too that we needed to escape quotes in the content.