Optimise prometheus_buckets:position/2
Examine the condition inlined in the helper instead of passing a
function object. Also ensure that the entry point has checked the types,
when looping on the inner function, the JIT compiler will know already
that the inputs are lists and numbers, skipping dynamic type checking on
every iteration. The `lists` module implements such optimisation
extensively.