Optimise handling of integers and floats for counters, gauges, summaries
The point is that these tables keep separate positions in the tuple for
integers and for floats, as on the integer position we can use the
optimised `ets:update_counter/3,4`, and probabilistically integers are
the most common input.

The issue is that under some circumstances, integer values would be
appended to the float position, which, while correct, misses the
opportunity for the optimisation.

Note too that gauges could take the value `undefined` to effectively
block changes until explicitly set. This was implemented, but not
tested, and in some edge-cases (the integer/float optimisation) it would
be buggy.