Fix rebar.config.script when the src dir doesn't exist
This can happen when compiling a project that uses this library as a
dependency and has no `src/` directory, arguably an edge case though.

The intend of the script is to prepare sections in ex_doc according to
the directories, so that metric modules are grouped into a metrics
section, etc.