Optimise histograms
Two things: when fetching for the buckets, lookup (and therefore only
copy) the bucket config instead of the entire record). Second, keep
buckets as a big tuple instead of a big list, which is smaller and
therefore cheaper to copy, and also can apply binary search during the
bucket lookup instead of linear search over a list.