prometheus_text_format: Avoid allocation in `has_special_char/1`
`prometheus_text_format:has_special_char/1` is called very often when
a registry contains many metrics with label pairs. We can use
`binary:match/2` to search within a label binary for the special
characters (newline, backslash and double-quote) without allocation.

The old code using binary match syntax creates a match context every
time the function is called (except, not recursion - then the match
context is reused). A match context allocates 5 words to the process
heap when it is created. When matching many many binaries this scales to
create a noticeable amount of short-lived garbage.

In comparison `binary:match/2` with a precompiled match pattern does not
allocate. The BIF for it is also very well optimized, using `memchr`
since OTP 22.