prometheus_text_format: Fix argument order for `format_into/3` fun
The formatting function should take the state as the first argument and
new data as the second. See `format/1` which passes in the trailing
linefeed as the second argument. This doesn't make much difference when
calling `format/1` because the swapped arguments only caused the metrics
to be concatenated in reverse. But this fix is important for using
`format_into/3` correctly, for example with `cowboy_req:stream_body/3`.